Move Strapline folder in its own folder
