+++ components/Strapline/index.tsx
@@ -1,0 +1,4 @@
+import Strapline from './Strapline';
+import type { IconChoice } from './types';
+
+export { Strapline as default, IconChoice };
