+++ components/Strapline/types.tsx
@@ -1,0 +1,8 @@
+type IconChoice = 'thumbs up' | 'thumbs down';
+
+type StraplineProps = {
+    icon: IconChoice;
+    text: string;
+};
+
+export type { IconChoice, StraplineProps };
