+++ pages/index.tsx
@@ -11,7 +11,7 @@ import type { Topics } from '../components/blocks/topics_block';
 import HeadingSlackCTA from '../components/HeadingSlackCTA';
 import FooterSlackCTA from '../components/FooterSlackCTA';
 import Footer from '../components/Footer';
-import Strapline from '../components/Strapline';
+import Strapline from '../components/Strapline/';
 import StreamField from '../components/StreamField/';
 
 const Home: NextPage<{ page: AreWeHeadlessYetHomePage; topics: Topics }> = ({
